+++ app/src/main/java/com/bd/deliverytiger/app/ui/dashboard/DashboardFragment.kt
@@ -579,7 +579,7 @@ class DashboardFragment : Fragment() {
 
     private fun getCourierUsersInformation() {
         viewModel.getCourierUsersInformation(SessionManager.courierUserId).observe(viewLifecycleOwner, Observer { model ->
-            SessionManager.collectionCharge = model.collectionCharge.toInt()
+            SessionManager.collectionCharge = model.collectionCharge
             SessionManager.merchantDistrict = model.districtId
             isQuickBookingEnable = model.isQuickOrderActive
             initRetentionManagerData(model?.adminUsers?.userId ?: 0, model?.adminUsers?.fullName ?: "", model?.adminUsers?.mobile ?: "")
